refactor: restructure to cmd/cli and pkg/ layout
