+++ Dockerfile.action
@@ -5,7 +5,7 @@ COPY go.mod go.sum ./
 RUN go mod download
 # .dockerignore excludes sensitive files (.env, .git, etc.)
 COPY . .  # NOSONAR
-RUN CGO_ENABLED=0 GOOS=linux go build -ldflags="-s -w" -o /heisenberg .
+RUN CGO_ENABLED=0 GOOS=linux go build -ldflags="-s -w" -o /heisenberg ./cmd/cli
 
 # Stage 2: Runtime with Playwright
 # Playwright requires root for browser installation; acceptable for GitHub Action containers
