+++ pkg/llm/client_test.go
