+++ pkg/llm/progress.go
