+++ pkg/llm/progress_test.go
