+++ pkg/trace/snapshot.go
