+++ pkg/trace/snapshot_test.go
