+++ pkg/trace/trace_test.go
